+++ .gitattributes
@@ -22,7 +22,7 @@
 /src/app/formoverview/formoverview-function.ts linguist-generated
 /src/app/home/home-function.ts linguist-generated
 /src/app/post-login/postlogin-function.ts linguist-generated
-/src/app/resbumit/resubmit-function.ts linguist-generated
+/src/app/resubmit/resubmit-function.ts linguist-generated
 /src/webapp/auth/auth-function.ts linguist-generated
 /src/webapp/login/login-function.ts linguist-generated
 /src/webapp/logout/logout-function.ts linguist-generated
