+++ package.json
@@ -10,8 +10,8 @@
     "bundle:app/home/home.lambda:watch": "npx projen bundle:app/home/home.lambda:watch",
     "bundle:app/post-login/postlogin.lambda": "npx projen bundle:app/post-login/postlogin.lambda",
     "bundle:app/post-login/postlogin.lambda:watch": "npx projen bundle:app/post-login/postlogin.lambda:watch",
-    "bundle:app/resbumit/resubmit.lambda": "npx projen bundle:app/resbumit/resubmit.lambda",
-    "bundle:app/resbumit/resubmit.lambda:watch": "npx projen bundle:app/resbumit/resubmit.lambda:watch",
+    "bundle:app/resubmit/resubmit.lambda": "npx projen bundle:app/resubmit/resubmit.lambda",
+    "bundle:app/resubmit/resubmit.lambda:watch": "npx projen bundle:app/resubmit/resubmit.lambda:watch",
     "bundle:webapp/auth/auth.lambda": "npx projen bundle:webapp/auth/auth.lambda",
     "bundle:webapp/auth/auth.lambda:watch": "npx projen bundle:webapp/auth/auth.lambda:watch",
     "bundle:webapp/login/login.lambda": "npx projen bundle:webapp/login/login.lambda",
